Griffin Nest
Made the rough shape of the nest for the griffins. I didn't go too into detail as there might be a better way to do these to get more quality for gradient info and shape in ZBrush. Possibly good to use for a low res model version though; We'll see.